tools/ignite/pyproject.toml
[project]
name = "ignite"
version = "2025.6.0"
description = "CLI tool to initialize and manage development environment configuration from a workspace.yml file"
requires-python = ">=3.9,<3.14"
dependencies = [
    "typer>=0.16.0",
    "pydantic>=2.11.7",
    "jsonschema>=4.22.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.1.1"
]

[project.scripts]
ignite = "ignite.main:cli"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
flake8-pyproject = "^1.2.3"
flake8-docstrings = "^1.7.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pytest-mock = "^3.14.0"
pytest-xdist = "^3.7.0"
assertpy = "^1.1"
coverage = "^7.3.2"

[build-system]
requires = ["poexy-core>=2025"]
build-backend = "poexy_core.api" 

[tool.poexy.binary]
name = "ignite"
entry_point = "ignite.main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]

[tool.coverage.run]
source = ["ignite"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/__pycache__/*",
    "*/.?*/*",
    "*/venv/*",
    "*/.venv/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError"
]

[tool.coverage.paths]
source = ["ignite"]

[tool.flake8]
max-line-length = 88
exclude = [
    ".git",
    "__pycache__",
    "dist"
]
extend-ignore = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D107",
    "D200",
    "D202",
    "D205",
    "D400",
    "D401",
]